+++ src/index.ts
@@ -1,9 +1,11 @@
 export { RetryManager } from './core/RetryManager';
-export { InMemoryRequestStore, type RequestStore } from './store/RequestStore';
+export { InMemoryRequestStore} from './store/InMemoryRequestStore';
 export {
   type RetryMode,
   type RetryHooks,
   type RetryManagerOptions,
   type RetryStrategy,
   type AxiosRetryerRequestConfig,
+  type RequestStore,
+  type RetryPlugin,
 } from './types';
