+++ src/store/InMemoryRequestStore.ts
@@ -1,25 +1,6 @@
 'use strict';
 
-import type { AxiosRetryerRequestConfig } from '../types';
-
-export interface RequestStore {
-  /**
-   * Add a request config to the store
-   * */
-  add(request: AxiosRetryerRequestConfig): void;
-  /**
-   * Remove a request config to the store
-   * */
-  remove(request: AxiosRetryerRequestConfig): void;
-  /**
-   * Get all request configs from the store
-   * */
-  getAll(): AxiosRetryerRequestConfig[];
-  /**
-   * Clear request store
-   * */
-  clear(): void;
-}
+import type {AxiosRetryerRequestConfig, RequestStore} from '../types';
 
 export class InMemoryRequestStore implements RequestStore {
   private requests: AxiosRetryerRequestConfig[] = [];
